Made simple loading screen and added music to main menu
