+++ Project Bergamot/Assets/Media/Music Asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 6a10353aa9000a54e937c2ffe7696042
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
