+++ Project Bergamot/Assets/Media/Music Asset/Free Game Menu Music Pack License.pdf.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: cc044dd9f92258d42a415589e1a00ea0
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
