+++ Project Bergamot/Assets/Media/Music Asset/Free Game Menu Music Pack.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 03db820a95b0b194fb033c850d7cf03c
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
