+++ Project Bergamot/Assets/Media/Music Asset/Free Game Menu Music Pack/11. Poseidon's Realm.wav.meta
@@ -1,0 +1,22 @@
+fileFormatVersion: 2
+guid: 33fb419c0ef254342acf6aa68d3dc8da
+AudioImporter:
+  externalObjects: {}
+  serializedVersion: 6
+  defaultSettings:
+    loadType: 0
+    sampleRateSetting: 0
+    sampleRateOverride: 44100
+    compressionFormat: 1
+    quality: 1
+    conversionMode: 0
+  platformSettingOverrides: {}
+  forceToMono: 0
+  normalize: 1
+  preloadAudioData: 1
+  loadInBackground: 0
+  ambisonic: 0
+  3D: 1
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
