+++ Project Bergamot/Assets/Media/Music Asset/Free Game Menu Music Pack/17. New Morning, New Sun.wav.meta
@@ -1,0 +1,22 @@
+fileFormatVersion: 2
+guid: 6655963f3f1df9647b67165beb8f477c
+AudioImporter:
+  externalObjects: {}
+  serializedVersion: 6
+  defaultSettings:
+    loadType: 0
+    sampleRateSetting: 0
+    sampleRateOverride: 44100
+    compressionFormat: 1
+    quality: 1
+    conversionMode: 0
+  platformSettingOverrides: {}
+  forceToMono: 0
+  normalize: 1
+  preloadAudioData: 1
+  loadInBackground: 0
+  ambisonic: 0
+  3D: 1
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
