+++ Project Bergamot/Assets/Media/Music Asset/Free Game Menu Music Pack/18. A Fool's Affair.wav.meta
@@ -1,0 +1,22 @@
+fileFormatVersion: 2
+guid: a8f8ba592f0aa89459093a950798d9c5
+AudioImporter:
+  externalObjects: {}
+  serializedVersion: 6
+  defaultSettings:
+    loadType: 0
+    sampleRateSetting: 0
+    sampleRateOverride: 44100
+    compressionFormat: 1
+    quality: 1
+    conversionMode: 0
+  platformSettingOverrides: {}
+  forceToMono: 0
+  normalize: 1
+  preloadAudioData: 1
+  loadInBackground: 0
+  ambisonic: 0
+  3D: 1
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
