+++ Project Bergamot/Assets/Media/Music Asset/Free Game Menu Music Pack/3. Funk Pump.wav.meta
@@ -1,0 +1,22 @@
+fileFormatVersion: 2
+guid: 1f03947c5ff14354e8bb7fe033ea14fb
+AudioImporter:
+  externalObjects: {}
+  serializedVersion: 6
+  defaultSettings:
+    loadType: 0
+    sampleRateSetting: 0
+    sampleRateOverride: 44100
+    compressionFormat: 1
+    quality: 1
+    conversionMode: 0
+  platformSettingOverrides: {}
+  forceToMono: 0
+  normalize: 1
+  preloadAudioData: 1
+  loadInBackground: 0
+  ambisonic: 0
+  3D: 1
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
