--- Project Bergamot/Assets/Prefabs/UI PREFABS/CarCell.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 20a78bf98cfcff14eb58aeb5b1a9a5c1
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
