--- Project Bergamot/Assets/Prefabs/UI PREFABS/CarCellObj.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: f4417219bdc03de478b882d4c301f2d1
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
