+++ Project Bergamot/Assets/Prefabs/UI PREFABS/MapUIPrefab.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 4ec341a4caa12334c809c46a6af2e475
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
