--- Project Bergamot/Assets/Prefabs/UI PREFABS/TestCarCellObj.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 924552a249ffe0b4385aca5dbf94b61b
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
