+++ Project Bergamot/Assets/Prefabs/UI PREFABS/testCharacterUi.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: bb261c79c84569e418ea90204efc6e4f
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
