+++ Project Bergamot/Assets/Prefabs/UI PREFABS/testCharacterUi/CarCellObj.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: f4417219bdc03de478b882d4c301f2d1
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
