+++ Project Bergamot/Assets/Prefabs/UI PREFABS/testCharacterUi/Player.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 91beb4d0e8f123543a883f9a92e3149d
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
