+++ Project Bergamot/Assets/Prefabs/UI PREFABS/testCharacterUi/PlayerCell.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 2bce5aa26486a674da501c6e748f43d2
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
