+++ Project Bergamot/Assets/Prefabs/UI PREFABS/testCharacterUi/TestCarCellObj.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 924552a249ffe0b4385aca5dbf94b61b
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
