+++ Project Bergamot/Assets/Scripts/UI/UiManagment.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: be2244256f13de24c9b5332d61069ddf
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
