+++ Project Bergamot/Assets/Scripts/UI/UiManagment/SceneSwitch.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: f4189287532f9cb4194cc4ce103cd3c0
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
