+++ Project Bergamot/Assets/Scripts/UI/map/mapManager.cs
@@ -6,10 +6,29 @@ using UnityEngine.UI;
 
 public class mapManager : MonoBehaviour
 {    
-    [SerializeField] private string mapName;
+    private string mapName;
+    public GameObject loadingScreen;
+    public Slider loadingSlider;
 
     public void mapSelected(string mapName)
     {
         SceneManager.LoadScene(mapName);
+    }
+
+    public void loadLevel(string name)
+    {
+        StartCoroutine(LoadAsync(name));
+    }
+
+    IEnumerator LoadAsync(string name)
+    {
+        AsyncOperation asyncOperation = SceneManager.LoadSceneAsync(name);
+        loadingScreen.SetActive(true);
+        while(!asyncOperation.isDone)
+        {
+           float loadingProgress = Mathf.Clamp01(asyncOperation.progress / .9f);
+           loadingSlider.value = loadingProgress;
+           yield return null;          
+        }
     }
 }
