--- Project Bergamot/Assets/prefabs/UI PREFABS/Player.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 91beb4d0e8f123543a883f9a92e3149d
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
