+++ Project Bergamot/UserSettings/EditorUserSettings.asset
@@ -27,10 +27,10 @@ EditorUserSettings:
       value: 22424703114646680e0b0227036c6a3956241b2f222d357f1e2c1136e1f43d37ecd435fbe23f31722c0ce6281d
       flags: 0
     RecentlyUsedScenePath-7:
-      value: 22424703114646680e0b0227036c6a3956241b2f222d357f00280d00e7ec313bf6ee39e7a92f31352d1b
+      value: 22424703114646680e0b0227036c6a3956241b2f222d357f01261c37ebee3376f7e93ffdfe
       flags: 0
     RecentlyUsedScenePath-8:
-      value: 22424703114646680e0b0227036c6a3956241b2f222d357f01261c37ebee3376f7e93ffdfe
+      value: 22424703114646680e0b0227036c6a3956241b2f222d357f00280d00e7ec313bf6ee39e7a92f31352d1b
       flags: 0
     RecentlyUsedScenePath-9:
       value: 22424703114646680e0b0227036c6a3956241b2f222d357f1e3d1c21f6cd3136f7a923e7ee2e26
